WIP - academic sessions

+++ database/migrations/2021_07_29_132747_add_academic_session_id_to_courses_table.php
@@ -14,7 +14,7 @@ class AddAcademicSessionIdToCoursesTable extends Migration
     public function up()
     {
         Schema::table('courses', function (Blueprint $table) {
-            $table->unsignedBigInteger('academic_session_id')->nullable();
+            $table->unsignedInteger('academic_session_id')->nullable();
             $table->foreign('academic_session_id')->references('id')->on('courses')->onDelete('cascade');
         });
     }
