+++ database/migrations/2021_07_29_135145_add_academic_session_id_to_disciplines_table.php
@@ -14,7 +14,7 @@ class AddAcademicSessionIdToDisciplinesTable extends Migration
     public function up()
     {
         Schema::table('disciplines', function (Blueprint $table) {
-            $table->unsignedBigInteger('academic_session_id')->nullable();
+            $table->unsignedInteger('academic_session_id')->nullable();
             $table->foreign('academic_session_id')->references('id')->on('users')->onDelete('cascade');
         });
     }
